convert all decimals to integers for conversion results from postgresql

+++ backend/db/models/spot.js
@@ -93,7 +93,7 @@ module.exports = (sequelize, DataTypes) => {
         },
       },
       lat: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.INTEGER,
         allowNull: false,
         validate: {
           withinRange(val) {
@@ -104,7 +104,7 @@ module.exports = (sequelize, DataTypes) => {
         },
       },
       lng: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.INTEGER,
         allowNull: false,
         validate: {
           withinRange(val) {
@@ -127,7 +127,7 @@ module.exports = (sequelize, DataTypes) => {
         validate: { len: [1, 300] },
       },
       price: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.INTEGER,
         allowNull: false,
         validate: {
           min: 0,
